fix(appsec): flask hangs on empty body requests
